[fusequery/query] move configs/clusters to src
