--- fusequery/clusters/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "fuse-query-clusters"
-version = "0.1.0"
-authors = ["Datafuse Authors <opensource@datafuselabs.com>"]
-license = "Apache-2.0"
-publish = false
-edition = "2018"
-
-[dependencies] # In alphabetical order
-# Workspace dependencies
-fuse-query-configs = {path = "../configs" }
-
-# Crates.io dependencies
-paste = "^1.0"
-serde = { version = "1.0", features = ["derive"] }
-snafu = "0.6.8"
-
-[dev-dependencies]
-pretty_assertions = "0.7"
