--- fusequery/clusters/src/lib.rs
@@ -1,15 +1,0 @@
-// Copyright 2020-2021 The Datafuse Authors.
-//
-// SPDX-License-Identifier: Apache-2.0.
-
-mod cluster_test;
-
-mod cluster;
-mod error;
-mod node;
-
-pub use fuse_query_configs as configs;
-
-pub use crate::cluster::{Cluster, ClusterRef};
-pub use crate::error::ClusterError;
-pub use crate::node::Node;
