--- fusequery/configs/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "fuse-query-configs"
-version = "0.1.0"
-authors = ["Datafuse Authors <opensource@datafuselabs.com>"]
-license = "Apache-2.0"
-publish = false
-edition = "2018"
-
-[dependencies] # In alphabetical order
-# Crates.io dependencies
-structopt = "0.3"
-num_cpus = "1.0"
