--- fusequery/metrics/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "fuse-query-metrics"
-version = "0.1.0"
-authors = ["Datafuse Authors <opensource@datafuselabs.com>"]
-license = "Apache-2.0"
-publish = false
-edition = "2018"
-
-[dependencies] # In alphabetical order
-# Workspace dependencies
-fuse-query-configs = {path = "../configs" }
-
-# Crates.io dependencies
-snafu = "0.6.8"
-metrics = "0.14.2"
-metrics-exporter-prometheus = "0.3.0"
