+++ fusequery/query/Cargo.toml
@@ -18,9 +18,6 @@ common-datavalues = {path = "../../common/datavalues"}
 common-functions = {path = "../../common/functions"}
 common-planners = {path = "../../common/planners"}
 
-fuse-query-clusters = {path = "../clusters" }
-fuse-query-configs = {path = "../configs" }
-fuse-query-metrics = {path = "../metrics"}
 
 # github dependencies
 arrow = { git = "https://github.com/apache/arrow", rev="46161d2" }
@@ -37,6 +34,7 @@ indexmap = "1.6.1"
 lazy_static = "1.4.0"
 log = "0.4"
 metrics = "0.14.2"
+metrics-exporter-prometheus = "0.3.0"
 msql-srv = "0.9"
 num = "0.4"
 num_cpus = "1.0"
@@ -48,6 +46,7 @@ serde = { version = "1.0", features = ["derive"] }
 simplelog = "0.9"
 snafu = "0.6.8"
 sqlparser = "0.8"
+structopt = "0.3"
 thiserror = "1.0"
 threadpool = "1.8.1"
 tokio = { version = "1.0", features = ["macros", "rt","rt-multi-thread", "sync"] }
