+++ fusequery/query/src/clusters/cluster_test.rs
@@ -3,10 +3,11 @@
 // SPDX-License-Identifier: Apache-2.0.
 
 #[test]
-fn test_cluster() -> crate::error::ClusterResult<()> {
+fn test_cluster() -> crate::error::FuseQueryResult<()> {
     use pretty_assertions::assert_eq;
 
-    use crate::*;
+    use crate::clusters::cluster::Cluster;
+    use crate::clusters::node::Node;
 
     let cluster = Cluster::empty();
