+++ fusequery/query/src/clusters/mod.rs
@@ -1,0 +1,11 @@
+// Copyright 2020-2021 The Datafuse Authors.
+//
+// SPDX-License-Identifier: Apache-2.0.
+
+mod cluster_test;
+
+mod cluster;
+mod node;
+
+pub use cluster::{Cluster, ClusterRef};
+pub use node::Node;
