+++ fusequery/query/src/metrics/metric_service.rs
@@ -2,10 +2,10 @@
 //
 // SPDX-License-Identifier: Apache-2.0.
 
-use fuse_query_configs::Config;
 use metrics_exporter_prometheus::PrometheusBuilder;
 
-use crate::error::{MetricError, MetricResult};
+use crate::configs::Config;
+use crate::error::{FuseQueryError, FuseQueryResult};
 
 pub struct MetricService {
     conf: Config,
@@ -16,7 +16,7 @@ impl MetricService {
         MetricService { conf }
     }
 
-    pub fn make_server(&self) -> MetricResult<()> {
+    pub fn make_server(&self) -> FuseQueryResult<()> {
         let addr = self
             .conf
             .metric_api_address
@@ -26,7 +26,7 @@ impl MetricService {
             .listen_address(addr)
             .install()
             .map_err(|e| {
-                MetricError::build_internal_error(format!(
+                FuseQueryError::build_internal_error(format!(
                     "Metrics prometheus exporter error: {:?}",
                     e
                 ))
