+++ fusequery/query/src/metrics/mod.rs
@@ -2,7 +2,6 @@
 //
 // SPDX-License-Identifier: Apache-2.0.
 
-mod error;
 mod metric_service;
 
 pub use metric_service::MetricService;
